Update FP_adder and right_shifter
The FP adder module is basically functional now, and it passes the current testbench. More corner cases are to be considered.

The right_shifter module contained a serious bug, in which shift_stg3 were assigned twice and shift_stg4 were not assigned. This caused its output to be high Z. It is fixed now.